Code reorganizing
