+++ lib/rbcli/configuration/config.rb
