+++ lib/rbcli/logging/logging.rb
